requirements.txt
djangorestframework
Pillow
psycopg2-binary
python-decouple
djangorestframework-simplejwt
celery
redis